Exported additional functions to a new c and h files of the decrypter.
